pyproject.toml
[project]
name = "claude-skill-tester"
version = "0.1.0"
description = "Test battery for evaluating Claude skill trigger accuracy"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
sdk = ["claude-agent-sdk>=0.1.50"]

[project.scripts]
skill-test = "skill_tester.__main__:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
